+++ aider/main.py
@@ -157,15 +157,15 @@ def main(argv=None, input=None, output=None, force_git_root=None):
         default=False,
         help="Override to skip model availability check (default: False)",
     )
-    default_4_turbo_model = models.GPT4_1106_PREVIEW
+    default_4_turbo_model = "gpt-4-1106-preview"
     core_group.add_argument(
         "--4-turbo",
         "--4turbo",
         "--4",
         action="store_const",
         dest="model",
-        const=default_4_turbo_model.name,
-        help=f"Use {default_4_turbo_model.name} model for the main chat (gpt-4 is better)",
+        const=default_4_turbo_model,
+        help=f"Use {default_4_turbo_model} model for the main chat (gpt-4 is better)",
     )
     default_3_model = models.GPT35_1106
     core_group.add_argument(